encapsulated starting bot out of index
